requirements.txt
pytorch
torchvision
torchaudio
opencv-python
tensorboard